git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11273 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/segment.h
@@ -284,7 +284,7 @@ class CSegment {
 		ubyte			m_function;
 		ubyte			m_flags;
 		ubyte			m_props;
-		sbyte			m_value;
+		short			m_value;
 		sbyte			m_nMatCen;
 		fix			m_xDamage [2];
 		fix			m_xAvgSegLight;
